requirements.txt
Django==5.2.6
djangorestframework==3.16.1
drf-spectacular==0.28.0
django-filter==25.1
Markdown==3.9
djangorestframework_simplejwt==5.5.1
dotenv == 0.9.9
Pillow == 11.3.0
psycopg2-binary==2.9.10
